dim: don't check everything in dim setup
Laurent got mightily confused about this, and assumed he's getting
volunteered as an intel maintainer because dim asks for an intel
remote.

With the branch autodetection we can postpone all the remote checks to
when we need them. This is similar to the changes to update-branches
to only update branches that exist, but for remotes.

Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>