drm-misc: Document small drivers expectations
For the experiement we have right now Eric (with vc4) and Sean Paul
(with rockchip and zte) volunteering, and Gerd (entire pile of qemu
drivers) and Boris (atmel) are also considering to participate. I
think that's enough to get started and figure things out as we go.

I tried to summarize the main points from the rfc discussions into a
short chapter.

v2: Spelling fixes (Anholt).

v3: Clarify which branches to use (Sean).

Cc: Boris Brezillon <boris.brezillon@free-electrons.com>
Cc: Eric Anholt <eric@anholt.net>
Cc: Sean Paul <seanpaul@chromium.org>
Cc: Gerd Hoffmann <kraxel@redhat.com>
Cc: Mark Yao <mark.yao@rock-chips.com>
Cc: Shawn Guo <shawnguo@kernel.org>
Acked-by: Sean Paul <seanpaul@chromium.org>
Acked-by: Gerd Hoffmann <kraxel@redhat.com>
Acked-by: Boris Brezillon <boris.brezillon@free-electrons.com>
Acked-by: Eric Anholt <eric@anholt.net>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>