dim: Add pull request tag template
Each pull request is accompanied by a summary that is stored in the git tag
from which it is generated. These summaries all share the same template with
headers classifying changes to UAPI, Cross-subsystem, Core, and Drivers. This
patch adds this template to the tag summary automatically in dim pull-request.

Changes in v2:
  - Tweaked the template var name s/PULL/TAG/ (Daniel)
Changes in v3:
  - Use git tag -F- to ingest template (Jani)
  - Tweak naming/comments again to hopefully clarify things (Jani)

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>