dim: Force-push in update-next
The autogenerated driver date update patch doesn't have a r-b tag for
obvious reasons, which gets caught up in the new r-b/a-b check in dim
push-branch. Hack around this.

Note this only overrides the dim checks, it's still a ff-only git push
(the -f for that needs to be added as a push-branch parameter).

Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>