qf: Move "gitk" from case to new subcommand.
No functional change. Only moving from the case
to its own function, with the arguments now shifted.

Notice that "shift" is not needed anymore since it is
getting shifted by default for all subcommands.

Also this is the last command to move from the case
block to its own function. So case block is not entirely
removed.

Cc: Jani Nikula <jani.nikula@intel.com>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Acked-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Acked-by: Jani Nikula <jani.nikula@intel.com>