dim.rst: Document aliases extension on dimrc.
On my own workflow I was missing a way to download mboxes
directly from patchwork with the patchwork id. So my first
reflex was to modify dim to fulfil my needs. However that
was increasing dim in complexity and dependencies and leaving
that messy.

That was when Jani suggested me the dimrc extension with the
example that is now part of this spec.

That was clean and simple enough to understand, so Daniel
suggested me to add it to the spec.

For record let's put my final local solution that lays now on
my own ~/.dimrc

dim_pwaq()
{
	if [ -n "$1" ]; then
		curl https://patchwork.freedesktop.org/patch/$1/mbox/ | dim_apply_queued
	else
	echo "Give me a patchwork id"
	fi
}

v2: Use code-block directive. Get's cleaner and make check happy.
v3: Use a generic block instead of sphinx directive otherwise
    rst2man can get confused on pur man generation.

Cc: Jani Nikula <jani.nikula@intel.com>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>