drm/i915: Apply Wa Display #1183 on skl, kbl, and cfl.
Wa Display #1183 was recently added to workaround
"Failures when enabling DPLL0 with eDP link rate 2.16
or 4.32 GHz and CD clock frequency 308.57 or 617.14 MHz
(CDCLK_CTL CD Frequency Select 10b or 11b) used in this
 enabling or in previous enabling."

However our code is already not following the
"Skylake Seqyences to Initialize Display" line by line
so it was really hard to map this workaround there.

The biggest difference is that Spec sequence expect that
by the time we are enabling DPLL0 we already know the
eDP link rate. What it is not true for us. We handle eDP
link rate as any other DP link rate at the modeset. With
only one small difference that we check a VCO when that
is available.

WARN: It seems that DPLL0 link rate was not designed to
change on the fly and we will probably need to find a more
robuts solution caching the eDP link rate somehow.

This Workaround was designed to minimize the impact only
to save the bad case with that link rates. But HW engineers
indicated that it should be safe to apply broadly. Although
they were expecting the DPLL0 link rate to be unchanged on
runtime.

Cc: Arthur J Runyan <arthur.j.runyan@intel.com>
Cc: Ville Syrjälä <ville.syrjala@linux.intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>