drm/i915: Add Display Gen info.
Introduce Display Gen. The goal is to use this to minimize
the amount of platform codename checks we have nowdays on
display code.

The introduction of a new platform should be just
gen >= current.

Just a gen++ without exposing any new feature or ip.
so this would minimize the amount of patches needed
for a bring-up specially holding them on internal branches.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>