xe/bo: Replace create_internal with a bitfield macro
We want to be able to specify more additional flags for internal buffers
than just GGTT so it makes sense to expose the whole bitfield.  At that
point, it makes sense to expose the whole bitfield.  Also, we may want
to allocate internal buffers in SMEM for the kernel for various things
like fence serials.