xe: Add a new xe_hw_fence, separate from everything else
This accomplishes two very useful things.  First, it abstracts the
concept of a fence that is triggered by HW writing a value and then
signaling an interrupt.  This is a general enough concept, it probably
doesn't even need to be Intel-specific.

Second, if fixes a reference counting loop where the drm_sched_job was
holding a reference to the fence returned by run_job but that was, in
tern, the reference count used for the xe_sched_job itself.  By making
the fence stand-alone, we can break the loop.