drm/xe: Start GGTT above WOPCM
The GuC owns the WOPCM space, thus we can't allocate GGTT address in
this area. Even though we likely configure the WOPCM to less than the
maximum value, to simplify the driver load (no need to fetch HuC + GuC
firmwares and determine there sizes before initializing the GGTT) just
start the GGTT allocation above the max WOPCM space. This might waste space
of GGTT space (WOPCM is 2MB on modern platforms) but we can live with
this.

Another benifit of this is the GuC bootrom can't access anything below
the WOPCM size so anything the bootom needs to access (e.g. a RSA key)
needs to be placed in the GGTT above the WOPCM maximum. Starting the
GGTT allocations above that give us that for free.

Signed-off-by: Matthew Brost <matthew.brost@intel.com>