drm/xe: Add GT abstraction layer
Not fully baked for multi-GT but this layer will help when go that way.

Follow up could move xe_gt.mem inline sub-structure to a structure /
layer in another file as likely either all the data in that structure is
unique or all of the data is shared with another tile (e.g. VRAM would
never be unique if GGTT is shared).

Signed-off-by: Matthew Brost <matthew.brost@intel.com>