drm/xe/WIP: Implement initial VM_BIND support.
For now, we only allow (un)binding a single memory entry with VM_BIND,
as we use the object lock to guard bo->vmas. We also pin the object when
bound, as we don't have support for evicting and bringing them back yet in
execbuf.

I upgraded the tests to handle the uapi changes, but for now we wait
unconditionally inside xe_vm.c for VM_(UN)BIND to complete. Afterwards
we can fix this.

XXX: We still specify a BO. However we also have xe_sync support,
should we get rid of the BO argument for anything but VM_BIND,
and use the sync obj locking in other cases? We can easily check if
obj we want to unbind is locked through acquire_ctx comparison.
This would allow us to unbind an entire range, even if external
objects are used.

Currently it's not allowed to touch or replace existing entries

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>