drm/xe: Don't communicate with the GuC to tear down kernel engines
Kernel engines are only torn down on driver unload which means the GuC
is getting blown away so no need to communicate with it. We can also
skip the step of routing the tear down through the DRM scheduler as this
isn't needed and tearing down of kernel contexts must be a blocking
operation to avoid the slabs being destroyed before the engines.

Signed-off-by: Matthew Brost <matthew.brost@intel.com>