drm/xe: Fix level 0 page table filled to end
xe_pt_idx(end, 0) would return 0 instead of 512. This caused wraparound,
and is the reason we used xe_pt_idx(end - 1, level) + 1 elsewhere. Do the
same for level 0, and unify the paths.

This makes the bind path for level 0 a simple debug print only,
and makes it clearer for higher levels how it works.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>