drm/xe: munmap style VM unbinds
e.g. Two binds are done to 0x0000-0x2000 and 0x2000-0x4000 and single
unbind is done to 0x1000-0x3000 which leaves us with mappings:
0x0000-0x1000 and 0x3000-0x4000

e.g. Two binds are done to 0x0000-0x2000 and 0x3000-0x5000 and single
unbind is done to 0x1000-0x4000 which leaves us with mappings:
0x0000-0x1000 and 0x4000-0x5000 (unbind are not an error if gaps in
unbinds range)

Internally a single VM unbind can be decomposed into multiple unbinds
and possibly 2 rebinds if the first / last VMA unbound is only partially
being umapped. This creates a window where a VMA the user expects to be
valid is invalid (between unbind and rebind). To work around this we
treat this like any other kernel op (eviction, userptr invalidation) and
use the dma-resv slots to ensure no jobs are inflight during this
window (or use preempt fences in the case of compute mode).

Why not just do a partial unbind on the edges? This gets tricky with
huge pages as we may no longer be able to use these large pages (e.g. 1
MB of an existing 2MB is being unmapped, now we need to switch to small
pages). For simplity, use the strategy detailed above.

Signed-off-by: Matthew Brost <matthew.brost@intel.com>