drm/xe: Use "_desc" to refer to static const platform data
Remove the TODO to embed the information in the xe_device: there is
already an anonymous struct holding that data. Instead rename the static
const data describing a platform to have a _desc suffix. That data is
pre-processed and the information that will be used in runtime, outside
initialization, is copied over to xe->info.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>