drm/xe: Initial GuC PC support with the GT sysfs as well
This patch enables the GuC SLPC freq management with all its defaults
features and using its defaults for all the platforms.

Let's get our frequency unlocked from the current static low freq
and allow GuC PC to take care of this for now.

This initial support doesn't include any API for tuning or controlling
the frequencies yet. This is coming in a follow-up work.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>