drm/xe: Make better use of pool of memory for user binds
We have a pool of memory in the migrate VM which is used to map bind VM
pages tables which need updating. Prior to this change this pool was
allocated in 1 page units while at most 64 bytes of memory is needed.

To be safe, allocate memory in 128 bytes units and reduce the pool size
by 64k.

Before this change we had 28 units available for user binds, after we
have 384 all while using 64k less memory.

Signed-off-by: Matthew Brost <matthew.brost@intel.com>