drm/xe: Introduce the CONFIG_DRM_XE_WERROR
Now that we have a clean build with the build of xe.mod.o
with the make M=drivers/gpu/drm/xe we can have the helper
of the compiler to void pushing any warning up.

And also without depending on i915 build which is disabled
in my xe environment.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>