drm/xe: Protect the freq management from races on gt_resets.
Protecting the user requested frequencies is not enough.
we need to also protect he shared min, max frequencies against
races.

So, let's use the big dumb lock to protect all the min max frequencies.

We avoid to shutdown PC when we are managing these frequencies and
we are really only ready after we ensure the user requested ones
got restored.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>