drm/xe: Add tuning registers
Tuning registers or "HW Recommended Settings" as mentioned in bspec,
were traditionally done in i915 together with the workarounds. That was
because the WA infra and WA list is one thing in i915. That is not the
case for xe.

Since those settings come from a different page in spec and that they
are not proper workarounds, keep them separate in a new xe_tuning.c
file.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>