drm/i915: Let's avoid using waitboost when SLPC is enabled.
SLPC already tune the gt frequencies properly depending on the
GuC scheduler and other metrics. We don't need to continue with
the waitboost mechanism.

However, due to the impact in the sysfs' boost api, we are
only enabling this for newer platforms that are still protected
by the force_probe. For the already released ones, waitboost
will continue to request slpc a certain frequency.

Cc: Ashutosh Dixit <ashutosh.dixit@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>