drm/xe/bo: Introduce a migrate interface
Introduce xe_bo_can_migrate() and xe_bo_migrate() functions. The former
checks whether migration will likely succeed and is intended to be used
as an early check as to whether migration attempts should be refused.
The latter will schedule a migration of a buffer object to a valid
placement under the reservation lock. The initial use-case will be dma-buf
migration in case the importer doesn't support p2p dma.

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>