drm/i915: Move some of the request decisions out of rps_boost function.
Ideally all the decisions should be made before calling the boost function.
And rps functions only receiving the rps struct. At least lets move most
of the decisions to the request component, but still leave the test
and set of the fence flag boost inside the rps because that might be time
sensitive.

Cc: Ashutosh Dixit <ashutosh.dixit@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>