drm/xe: Remove the xe_bo_populate() interface
Replace with xe_bo_validate() where needed. Difference is that
xe_bo_validate() is interruptible, and we need to watch out for
callers not yet being able to handle that.
In come cases, the call to xe_bo_populate() is removed, and in those
cases, the buffer object was already pinned in the relevant memory
type.
An interface like xe_bo_populate() should only be needed if we bind
directly to system memory; otherwise TTM will populate before moving
to another memory type.

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>