drm/xe: Introduce xe_pcode component.
The goal is that this component will be solely responsible for
any communication to PCODE. It will export functions to perform
specific actions rather than exposing generic read and write
operations.

This one starts with the PCODE min freq qos setup table needed
for the gen12 integrated parts.

There's still a lot to be done here in this component:
- documentation
- locks
- init done timeout for dgfx
- hw metrics
- display operations

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>