drm/xe: Add forcewake_all debugfs file.
We currently have no way to forcewake the gpu from the user space.
So, IGT gets kind of limited.

For instance, for the latency tests we need to directly
access the timestamp registers from IGT. If the gpu is in
RC6 for instance, these registers will return 0 and mess
with the tests. So, we need a reliable way from the IGT
to enforce the GT is awake.

This mechanism is similar to the i915_forcewake_user.
But here calling forcewake_all because we are really
going through all the GTs and using the forcewake_all
domain.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>