drm/xe: Delete xe_lrc_assert_ring_space
The WARN for ring space is bogus as the LRC head value can be stale
relative to job signaling and the ring head value. This results in the
DRM scheduler submitting jobs as it knows there is space in the ring via
job signaling while the LRC reads a stale value which indicates the ring
does not have space. Delete this assert.

Signed-off-by: Matthew Brost <matthew.brost@intel.com>