drm/xe/migrate: Fix and clarify some size calculations
We were using the curs->remaining rather than the curs->size field
to calculate the remaing size of the current vram segment. That
might blow up if the bo vram allocation is fragmented. Fix that.

Also update the batch buffer calculation slightly to limit the
blit size earlier and presuming that the batch buffer size calculation
is now correct and the workaround for issue #52 is no longer needed.

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>