drm/xe: Pagetable update
Use the gpu page-table walker to rewrite page-table building for bind.
There are a couple of problems we want to address by this:

1) For 64K alignment systems, We need the GEN12_PTE_PS64 bits to be able
to mix 64K-requiring VRAM and 4K-requiring system in the same 2MB L0
pagetable, which our uapi allows.

2) USM leaves need to be built at bind commit time, since they need to
move under the userptr invalidation lock to protect against racing with
the userptr invalidation mmu notifier. Moving under the lock will be
addressed as a follow up, but leaf building is moved to bind commit and
also uses the page-table walker for simplicity.

3) Avoid rewinding and restarting resource cursors.

4) Add documentation.

The patch is still missing some documentation and debug printouts, but
should be fully functional at this point. Has seen some rewrite since last
test on DG2, though. That will be fixed before merging.

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>