drm/xe/gt: Read topology from fuse registers
Reading GT topology (DSS masks and EU masks) from fuse registers is
a pre-requisite for several items:
 - Proper multicast/replicated (MCR) register handling
 - Some GT workarounds
 - Proper userspace operation

The topology masks are provided by hardware via one or more 32-bit
registers.  The DSS masks for future platforms are expected to
eventually grow beyond the storage of a u64, so we provide a custom mask
type based on the linux/bitmap.h facilities.  EU masks (i.e., the mask
of EUs available to a single DSS) are not expected to change
significantly in the foreseeable future, so we stick with a simpler u64
there.

v2:
 - Switch EU mask to its own bitmap-based type.  In practice we never
   expect there to be more than 32 bits to the mask, so bitmap.h's macro
   magic will treat this as a single 'unsigned long' anyway.  (Rodrigo)
 - Move register definitions into the .c for now; exactly how register
   defs should be handled in Xe is still being determined, but these
   specific registers shouldn't be needed in any other file.  (Rodrigo)

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>