drm/xe/migrate: Sync partial copies and clears.
If we hit an error while copying or clearing a large object, we might
already have submitted a partial copy- or clear operation on the object.

When the error is propagated to TTM it will immediately free the new
ttm resource we have submitted a gpu operation on, and unfortunately
there's no way to attach a fence to that resource, and that may lead
to the gpu attempting to touch freed system memory, and without an IOMMU
to stop it, it will.

So sync any partial migrate gpu operation if we hit an error.

This will partly fix issue #145

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>