drm/xe: Rework the xe_migrate_update_pgtables callbacks slightly
Use an ops struct rather than passing the function pointer directly,
in order for us to easily be able to add additional callback functions,
since later in the series we will be adding a callback to lock the
userptr notifier lock around submission.
Also use an embeddable base-class-like struct rather than a void
argument to pass customized parameters to the callbacks.

There is no functional change with this commit.

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>