drm/xe: Perform userptr pagetable updates under the notifier_lock
In order to protect the usm leaves from concurrent writing (during bind-
and unbind commit) and reading (during userptr invalidation), protect the
bind / unbind commit with the notifier lock in read mode.

Userptr invalidation in fault mode is made to wait also for the pagetable
update gpu writes which makes sure that they don't race.

If, during bind we detect a race with the userptr mmu notifier, we abort
and return -EAGAIN. If the same happens during unbind, we just wait for
everybody to exit the mmu notifier before proceeding with the unbind.

Since we now may attempt to free bos under the userptr lock which is
reclaim tainted and a spinlock, we need to use the deferred bo freeing
mechanism and free those bos when we release the notifier lock.

Finally add a config option DRM_XE_USERPTR_INVAL_INJECT that tests
the -EINVAL invalidation error path for bind operation. Currently that
causes issues with gpu page-table writes and xe_exec_basic, it looks
like issues with fence freeing when scheduler jobs are freed without
being submitted, so for now push out fixing this until we can insert
fences into the timeline at xe_sched_job_arm() time.

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>