FIXME: drm/i915/display: Remaining changes to make xe compile
[Planes, bos, fb, fbdev, fbc, tiling, stolen, frontbuffer. Still needs
to be chopped up to smaller pieces.]

Xe, the new Intel GPU driver, will re-use the i915 display.

At least for now, the plan is to use symbolic links and
adjust the build so we are building the display either for
i915 or for xe.

The display can be split out if needed.
Also the compilation is optional at this time

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>