drm/xe/guc_pc: RPe from PCODE precedes fused min freq.
GuC PC is in total control of the freq request. Then
the PCODE is the ultimate decided for the frequency,
regardless of what i915 requested or what guc has as its
defaults per platform.

So, although this is an unlikely scenario, it can happen
as we saw on DG2. Let's not fight with that, but accept
the PCODE's decisions.

This decision is prune to change. So let's keep updating it.
And move our rpn to the lowest possible found frequency.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>