drm/xe: Rework preempt fences handling
In order to facilitate a cleaner error handling in the rebind worker,
in particular with respect to early userptr invalidation at bind time,
split preempt fence handling up in alloc, arm and install, with the
idea that arm can be placed under the userptr notifier lock so that
we can easily abort and rerun the rebind worker if we hit a userptr
invalidation just before arm.

Also add some kerneldoc for the new preempt fence interfaces.

This is a prerequisite for a bigger additional userptr update.

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>