drm/xe/bo: Wait for bo bind in xe_bo_trigger_rebind() in fault mode
Before zapping the GPU ptes for a bo in bind mode, wait for an ongoing
bind or unbind so that we don't race when updating the ptes.

New pte update commits are blocked by the bo resv being held, so once
ongoning binds and unbinds are complete, zapping should be safe.

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>